test_timeOutEnvironment in separate file
